requirements.txt
Flask>=2.0.0
requests>=2.26.0
beautifulsoup4>=4.10.0
Pillow>=9.0.0
urllib3>=1.26.0
setuptools>=58.0.4
Werkzeug>=2.0.0